+++ car/src/main/java/pt/ulisboa/tecnico/softeng/car/services/remote/dataobjects/RestInvoiceData.java
@@ -4,22 +4,23 @@ import org.joda.time.DateTime;
 import org.joda.time.LocalDate;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
+import pt.ulisboa.tecnico.softeng.car.domain.RentACar;
 
 public class RestInvoiceData {
 	private String sellerNif;
 	private String buyerNif;
 	private String itemType;
-	private Double value;
+	private long value;
 	@JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd")
 	private LocalDate date;
-	private Double iva;
+	private long iva;
 	@JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm:ss.SSS")
 	private DateTime time;
 
 	public RestInvoiceData() {
 	}
 
-	public RestInvoiceData(String sellerNif, String buyerNif, String itemType, Double value, LocalDate date,
+	public RestInvoiceData(String sellerNif, String buyerNif, String itemType, long value, LocalDate date,
 			DateTime time) {
 		this.sellerNif = sellerNif;
 		this.buyerNif = buyerNif;
@@ -53,11 +54,11 @@ public class RestInvoiceData {
 		this.itemType = itemType;
 	}
 
-	public Double getValue() {
+	public long getValue() {
 		return this.value;
 	}
 
-	public void setValue(Double value) {
+	public void setValue(long value) {
 		this.value = value;
 	}
 
@@ -69,11 +70,11 @@ public class RestInvoiceData {
 		this.date = date;
 	}
 
-	public Double getIva() {
+	public long getIva() {
 		return this.iva;
 	}
 
-	public void setIva(Double iva) {
+	public void setIva(long iva) {
 		this.iva = iva;
 	}
